+++ src/BookmarksList.h
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2020-2022 Tobias Leupold <tl at stonemx dot de>
+// SPDX-FileCopyrightText: 2020-2024 Tobias Leupold <tl at stonemx dot de>
 //
 // SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-KDE-Accepted-GPL
 
@@ -41,9 +41,11 @@ private: // Structs
         bool okay;
     };
 
+public Q_SLOTS:
+    void newBookmark();
+
 private Q_SLOTS:
     void showContextMenu(const QPoint &point);
-    void newBookmark();
     void newManualBookmark();
     void renameBookmark();
     void editCoordinates();