+++ src/CoordinatesFormatter.cpp
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2020 Tobias Leupold <tl at stonemx dot de>
+// SPDX-FileCopyrightText: 2020-2024 Tobias Leupold <tl at stonemx dot de>
 //
 // SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-KDE-Accepted-GPL
 
@@ -6,6 +6,7 @@
 #include "CoordinatesFormatter.h"
 #include "KGeoTag.h"
 #include "Coordinates.h"
+#include "Settings.h"
 
 // KDE includes
 #include <KLocalizedString>
@@ -16,20 +17,66 @@
 // C++ includes
 #include <cmath>
 
-CoordinatesFormatter::CoordinatesFormatter(QObject *parent, QLocale *locale)
+CoordinatesFormatter::CoordinatesFormatter(QObject *parent, QLocale *locale, Settings *settings)
     : QObject(parent),
-      m_locale(locale)
+      m_locale(locale),
+      m_settings(settings)
 {
 }
 
 QString CoordinatesFormatter::formatLonLat(double value) const
 {
-    return m_locale->toString(std::abs(value), 'f', KGeoTag::degreesPrecision);
+    switch (m_settings->coordinatesFlavor()) {
+    case KGeoTag::DecimalDegrees:
+        return i18nc("Formatted coordinates as decimal degrees", "%1°",
+                     m_locale->toString(std::abs(value), 'f', KGeoTag::degreesPrecision));
+
+    case KGeoTag::DegreesDecimalMinutes:
+    {
+        double decimals;
+        double degrees;
+        double decimalMinutes;
+        decimals = std::modf(value, &degrees);
+        decimalMinutes = decimals * 60.0;
+        return i18nc("Formatted coordinates as degrees and decimal minutes", "%1° %2'",
+                     QString::number(std::abs(degrees)),
+                     m_locale->toString(std::abs(decimalMinutes), 'f', KGeoTag::minutesPrecision));
+    }
+
+    case KGeoTag::DegreesMinutesDecimalSeconds:
+    {
+        double decimals;
+        double degrees;
+        double decimalMinutes;
+        double minutes;
+        double decimalSeconds;
+        decimals = std::modf(value, &degrees);
+        decimalMinutes = decimals * 60.0;
+        decimals = std::modf(decimalMinutes, &minutes);
+        decimalSeconds = decimals * 60.0;
+        return i18nc("Formatted coordinates as degrees, minutes and decimal seconds",
+                     "%1° %2' %3\"",
+                     QString::number(std::abs(degrees)),
+                     QString::number(std::abs(minutes)),
+                     m_locale->toString(std::abs(decimalSeconds), 'f', KGeoTag::secondsPrecision));
+    }
+
+    }
+
+    // We can't reach here
+    return QString();
+}
+
+QString CoordinatesFormatter::format(const Coordinates &coordinates) const
+{
+    return i18nc("Formatted coordinates, \"lon, lat\" or \"lat, lon\"", "%1, %2",
+                 m_settings->latBeforeLon() ? lat(coordinates) : lon(coordinates),
+                 m_settings->latBeforeLon() ? lon(coordinates) : lat(coordinates));
 }
 
 QString CoordinatesFormatter::lon(const Coordinates &coordinates) const
 {
-    return i18nc("Formatted longitude with a cardinal direction", "%1° %2",
+    return i18nc("Formatted latitude or longitude with a cardinal direction", "%1 %2",
                  formatLonLat(coordinates.lon()),
                  coordinates.lon() >= 0 ? i18nc("Abbreviated cardinal direction \"East\"", "E")
                                         : i18nc("Abbreviated cardinal direction \"West\"", "W"));
@@ -37,7 +84,7 @@ QString CoordinatesFormatter::lon(const Coordinates &coordinates) const
 
 QString CoordinatesFormatter::lat(const Coordinates &coordinates) const
 {
-    return i18nc("Formatted latitude with a cardinal direction", "%1° %2",
+    return i18nc("Formatted latitude or longitude with a cardinal direction", "%1 %2",
                  formatLonLat(coordinates.lat()),
                  coordinates.lat() >= 0 ? i18nc("Abbreviated cardinal direction \"North\"", "N")
                                         : i18nc("Abbreviated cardinal direction \"South\"", "S"));
@@ -45,5 +92,6 @@ QString CoordinatesFormatter::lat(const Coordinates &coordinates) const
 
 QString CoordinatesFormatter::alt(const Coordinates &coordinates) const
 {
-    return m_locale->toString(coordinates.alt(), 'f', KGeoTag::altitudePrecision);
+    return i18nc("Formatted altitude in meters", "%1 m",
+                 m_locale->toString(coordinates.alt(), 'f', KGeoTag::altitudePrecision));
 }