+++ src/Settings.cpp
@@ -20,6 +20,21 @@ static const QLatin1String s_windowState("windowState");
 static const QLatin1String s_lastOpenPath("lastOpenPath");
 static const QLatin1String s_splitImagesList("splitImagesList");
 
+// Coordinates
+static const QLatin1String s_coordinates("coordinates");
+static const QLatin1String s_latBeforeLon("latBeforeLon");
+static const QLatin1String s_coordinatesFlavor("flavor");
+static const QHash<KGeoTag::CoordinatesFlavor, QString> s_flavorEnumToString {
+    { KGeoTag::DecimalDegrees,               QStringLiteral("DecimalDegrees") },
+    { KGeoTag::DegreesDecimalMinutes,        QStringLiteral("DegreesDecimalMinutes") },
+    { KGeoTag::DegreesMinutesDecimalSeconds, QStringLiteral("DegreesMinutesDecimalSeconds") }
+};
+static const QHash<QString, KGeoTag::CoordinatesFlavor> s_flavorStringToEnum {
+    { QStringLiteral("DecimalDegrees"),               KGeoTag::DecimalDegrees },
+    { QStringLiteral("DegreesDecimalMinutes"),        KGeoTag::DegreesDecimalMinutes },
+    { QStringLiteral("DegreesMinutesDecimalSeconds"), KGeoTag::DegreesMinutesDecimalSeconds }
+};
+
 // Map
 static const QLatin1String s_map("map");
 static const QLatin1String s_showCrosshairs("showCrosshairs");
@@ -125,6 +140,17 @@ static const QString s_bookmarksDataAlt = QStringLiteral("alt");
 Settings::Settings(QObject *parent) : QObject(parent)
 {
     m_config = KSharedConfig::openConfig();
+
+    // Initialize the "lat before lon" bool
+    auto group = m_config->group(s_coordinates);
+    m_latBeforeLon = group.readEntry(s_latBeforeLon, false);
+
+    // Initialize the coordinates flavor
+    const auto flavor = group.readEntry(s_coordinatesFlavor,
+                                        s_flavorEnumToString.value(KGeoTag::DecimalDegrees));
+    m_coordinatesFlavor = s_flavorStringToEnum.contains(flavor)
+                              ? s_flavorStringToEnum.value(flavor)
+                              : KGeoTag::DecimalDegrees;
 }
 
 // KMainWindow settings
@@ -520,4 +546,30 @@ bool Settings::selectNextUntagged() const
 {
     auto group = m_config->group(s_assignment);
     return group.readEntry(s_selectNextUntagged, true);
+}
+
+void Settings::saveLatBeforeLon(bool state)
+{
+    m_latBeforeLon = state;
+    auto group = m_config->group(s_coordinates);
+    group.writeEntry(s_latBeforeLon, state);
+    group.sync();
+}
+
+bool Settings::latBeforeLon() const
+{
+    return m_latBeforeLon;
+}
+
+void Settings::saveCoordinatesFlavor(KGeoTag::CoordinatesFlavor flavor)
+{
+    m_coordinatesFlavor = flavor;
+    auto group = m_config->group(s_coordinates);
+    group.writeEntry(s_coordinatesFlavor, s_flavorEnumToString.value(flavor));
+    group.sync();
+}
+
+KGeoTag::CoordinatesFlavor Settings::coordinatesFlavor() const
+{
+    return m_coordinatesFlavor;
 }