+++ src/SettingsDialog.h
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2020-2022 Tobias Leupold <tl at stonemx dot de>
+// SPDX-FileCopyrightText: 2020-2024 Tobias Leupold <tl at stonemx dot de>
 //
 // SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-KDE-Accepted-GPL
 
@@ -40,6 +40,9 @@ private: // Functions
 
 private: // Variables
     Settings *m_settings;
+
+    QComboBox *m_coordinatesOrder;
+    QComboBox *m_coordinatesFlavor;
 
     QComboBox *m_imageListsMode;
     QCheckBox *m_splitImagesList;