+++ CHANGELOG.rst
@@ -54,6 +54,12 @@ Added
   keyboard shortcut. Optionally, you can also have the next untagged image selected automatically
   afterwards. Thanks to Robert Menger for the respective feature request :-)
 
+* Added more functionality to the map's context menu: Bookmarks can now be added for the current
+  map center and it's coordinates can be copied to the clipboard.
+
+* The displayed order of coordinates can now be configured, either "Longitude, latitude" (as
+  before) or "Latitude, longitude".
+
 Changed
 =======